+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed an issue where empty assessment sets couldn't be created by `make_splits()` (#188).
 
+* A `fingerprint()` function was added to create a hash value that can be used to compare `rset` objects.
+
 # rsample 0.0.8
 
 * New `manual_rset()` for constructing rset objects manually from custom rsplits (tidymodels/tune#273).